% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_damerau_levenshtein.R
\name{calculate_damerau_levenshtein}
\alias{calculate_damerau_levenshtein}
\title{Calculate the Damerau-Levenshtein distance}
\usage{
calculate_damerau_levenshtein(target_col, response_col)
}
\arguments{
\item{target_col}{A string representing the target word to compare.}

\item{response_col}{A string representing the response word to compare against the target.}
}
\value{
A numeric value representing the Damerau-Levenshtein distance between `target_col` and `response_col`. A higher value indicates a greater number of edits or transpositions needed to transform one string into the other.
}
\description{
Computes the Damerau-Levenshtein distance between two strings located in `target_col` and `response_col`. The Damerau-Levenshtein distance is a metric that measures the number of single-character edits (insertions, deletions, substitutions) and transpositions required to change one string into the other.
}
\details{
This function uses the `stringdist` package's implementation of the Damerau-Levenshtein distance. This distance metric extends the Levenshtein distance by also considering transpositions of adjacent characters, making it useful for measuring similarity in cases where adjacent character transpositions are common.
}
\examples{
calculate_damerau_levenshtein("monkey", "monye")
calculate_damerau_levenshtein("article", "artim")

}
