% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jaro_winkler.R
\name{calculate_jaro_winkler}
\alias{calculate_jaro_winkler}
\title{Calculate the Jaro-Winkler distance}
\usage{
calculate_jaro_winkler(target_col, response_col)
}
\arguments{
\item{target_col}{A string representing the target word to compare.}

\item{response_col}{A string representing the response word to compare against the target.}
}
\value{
A numeric value representing the Jaro-Winkler distance between `target_col` and `response_col`. The distance ranges from 0 to 1, where a value closer to 1 indicates a higher similarity between the strings.
}
\description{
Computes the Jaro-Winkler distance between two strings located in `target_col` and `response_col`. The Jaro-Winkler distance measures the similarity between two strings by considering the number of matching characters and transpositions, with a prefix scale that gives more weight to strings that match from the beginning.
}
\details{
This function uses the `stringdist` package's implementation of the Jaro-Winkler distance. This metric extends the Jaro distance by giving more importance to matches at the beginning of the strings, which is useful for matching names and other short strings where early matches are often more significant.
}
\examples{
calculate_jaro_winkler("monkey", "monye")
calculate_jaro_winkler("article", "artim")

}
