% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_levenshtein.R
\name{calculate_levenshtein}
\alias{calculate_levenshtein}
\title{Calculate the Levenshtein distance}
\usage{
calculate_levenshtein(target_col, response_col)
}
\arguments{
\item{target_col}{A string representing the target word to compare.}

\item{response_col}{A string representing the response word to compare against the target.}
}
\value{
A numeric value representing the Levenshtein distance between `target_col` and `response_col`. A higher value indicates a greater number of edits needed to transform one string into the other.
}
\description{
Computes the Levenshtein distance between two strings located in `target_col` and `response_col`. The Levenshtein distance measures the number of single-character edits (insertions, deletions, or substitutions) required to change one string into the other.
}
\details{
This function utilizes the `stringdist` package's implementation of the Levenshtein distance. The distance is a common metric for evaluating the similarity between two strings, where a higher distance indicates more differences between the strings.
}
\examples{
calculate_levenshtein("monkey", "monye")
calculate_levenshtein("article", "artim")

}
