% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_errors.R
\name{classify_errors}
\alias{classify_errors}
\title{Classify Responses Based on Multiple Criteria}
\usage{
classify_errors(
  dataframe,
  access_col = "accessed",
  RA_col = "RA",
  cosine_limit_value = 0.46,
  response_col = "response_col",
  classify_RAs = FALSE
)
}
\arguments{
\item{dataframe}{A data frame containing the response data to be classified.}

\item{access_col}{A string specifying the name of the column in the dataframe that indicates whether the response was accessed (default is "accessed").}

\item{RA_col}{A string specifying the name of the column in the dataframe that indicates whether the response was a related attempt (default is "RA").}

\item{cosine_limit_value}{A numeric value specifying the threshold for cosine similarity used in classify_errors (default is 0.46).}

\item{response_col}{A string specifying the name of the column in the dataframe that contains the response status (default is "response_col").}

\item{classify_RAs}{A logical value indicating whether to classify rows with `RA` equal to 1 according to the conditions (default is FALSE). If FALSE, such rows will have the classify_errors results set to NA.}
}
\value{
A data frame with additional columns for each classify_errors category: `lexicalization`, `nonword`, `neologism`, `formal`, `unrelated`, `morphological`, `mixed`, and `semantic`. Each column contains binary values indicating the classify_errors status of each response, along with `no_response` and `human_check` columns.
}
\description{
This function classifies responses in a dataframe into different categories based on specified criteria. It adds new columns to the dataframe for each classify_errors category, including lexicalization, nonword, neologism, formal, unrelated, morphological, mixed, and semantic. Additionally, it computes a `human_check` column indicating if manual review is required based on the sum of classify_errors indicators and the `correct` column.
}
\details{
The function processes the dataframe by renaming columns and applying a series of conditions to create new classify_errors columns. The conditions are as follows:
- `lexicalization`: Identifies responses that are lexicalized based on specific conditions.
- `nonword`: Identifies responses that are classified as nonwords.
- `neologism`: Identifies responses that are classified as neologisms.
- `formal`: Identifies responses that are classified as formal.
- `unrelated`: Identifies responses that are classified as unrelated.
- `mixed`: Identifies responses that are classified as mixed.
- `semantic`: Identifies responses that are classified as semantic.

Each condition uses logical operations to check the values in the dataframe's columns and applies criteria based on those values to determine the classify_errors. The resulting classify_errorss are added as new columns in the dataframe with binary indicators (1 for true, 0 for false).

Additionally, the function includes:
- `no_response`: A binary column indicating if the `response_col` is "Response" and contains NA values (1 if TRUE, 0 if FALSE).
- `human_check`: A column indicating "required" if the sum of the classify_errors columns (from `lexicalization` to `no_response`) plus the `correct` column is 0 or greater than or equal to 2 for that row; otherwise, it is empty.

When `classify_RAs` is FALSE, rows where `RA` is 1 will have the classify_errors results set to NA. When `classify_RAs` is TRUE, classify_errors is performed normally regardless of the `RA` value.
}
