% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ismd.R
\name{count_ismd}
\alias{count_ismd}
\title{Count character transformations}
\usage{
count_ismd(string)
}
\arguments{
\item{string}{A string representing the sequence of transformations, typically derived from the output of the `similarity_vector` function.}
}
\value{
A data frame with counts of each transformation type: 'I' for insertions, 'S' for substitutions, 'M' for matches, and 'D' for deletions.
}
\description{
Counts the number of specific character transformations in a given string. The transformations are identified as "insertion" (i), "substitution" (s), "match" (m), and "deletion" (d). This function processes the string to tally the occurrences of each transformation type.
}
\details{
The function converts the input string to lowercase for case-insensitive counting. It then iterates through each character in the string, counting occurrences of 'i', 's', 'm', and 'd', which represent insertion, substitution, match, and deletion transformations, respectively.
}
