% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_char_count.R
\name{diff_char_count}
\alias{diff_char_count}
\title{Calculate the difference in amount of characters}
\usage{
diff_char_count(target_col, response_col)
}
\arguments{
\item{target_col}{A string representing the target word to compare.}

\item{response_col}{A string representing the response word to compare against the target.}
}
\value{
A numeric value representing the difference in phoneme count between `target_col` and `response_col`, as measured by the LCS method. A higher value indicates a greater difference in phoneme count.
}
\description{
Computes the difference in phoneme count between the two words words located in `target_col` and `response_col`.
}
\details{
The function uses the LCS method from the `stringdist` package to calculate the difference between the two strings. The LCS method determines how similar the two strings are by finding the length of their longest common subsequence and then calculates the difference based on this measure.
}
\examples{
diff_char_count("monkey", "monye")
diff_char_count("article", "artim")

}
