% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formal_similarity.R
\name{get_formal_similarity}
\alias{get_formal_similarity}
\title{Compute Formal Similarity Indexes Between Target and Response Columns}
\usage{
get_formal_similarity(
  dataframe,
  target_col = "item",
  response_col = "response",
  item_type = "task_type",
  source1 = NULL,
  source2 = NULL,
  source3 = NULL
)
}
\arguments{
\item{dataframe}{A dataframe containing the data to analyze.}

\item{target_col}{A string specifying the name of the column containing the target items. Default is `"item"`.}

\item{response_col}{A string specifying the name of the column containing the response items. Default is `"response"`.}

\item{item_type}{A string specifying the name of the column indicating the item type. Default is `"task_type"`.}

\item{source1}{An optional vector of words to check against the responses. Default is `NULL`.}

\item{source2}{An optional vector of words to check against the responses. Default is `NULL`.}

\item{source3}{An optional vector of words to check against the responses. Default is `NULL`.}
}
\value{
A dataframe with additional columns for the calculated similarity indexes, including:
\itemize{
  \item `shared_proportion`: Proportion of shared characters.
  \item `shared1char`: Shared starting letter indicator.
  \item `is_plural`: Plurality agreement indicator.
  \item `is_target_nonword`: Placeholder column.
  \item `same_root`: Root similarity indicator.
  \item `is_response_word`: Indicator if the response is found in the combined source lists.
}
}
\description{
This function computes various formal similarity indexes between the target and response columns in a dataframe. It calculates metrics such as shared proportion of characters, shared starting characters, plural/singular agreement, and root similarity. Additionally, it checks if the responses are present in specified source lists and adds this information to the dataframe.
}
\details{
The function calculates the following similarity metrics:
\itemize{
  \item `shared_proportion`: The proportion of shared characters between target and response.
  \item `shared1char`: A measure indicating whether the target and response share the same starting letter.
  \item `is_plural`: An indicator of whether the target and response agree in terms of plurality.
  \item `is_target_nonword`: (Placeholder for future implementation) Checks if the target is a non-word.
  \item `same_root`: A measure indicating whether the target and response share the same root.
}

The function also combines the provided source lists into a single unique list, and adds a new column, `is_response_word`, indicating whether each response is present in this combined list.
}
