% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_semantic_similarity.R
\name{get_semantic_similarity}
\alias{get_semantic_similarity}
\title{Calculate Cosine Similarity Between Words in a DataFrame}
\usage{
get_semantic_similarity(df, target_col, response_col, model)
}
\arguments{
\item{df}{A data frame containing the words to compare.}

\item{target_col}{A string specifying the name of the column in the dataframe that contains the target words.}

\item{response_col}{A string specifying the name of the column in the dataframe that contains the response words.}

\item{model}{A word embedding model used to generate word vectors. The model should have a `predict` method that can return word embeddings.}
}
\value{
A data frame with an additional column named `cosine_similarity`. This column contains the cosine similarity scores between the words in `target_col` and `response_col`. If an error occurs or the words are invalid, the value will be `NA`.
}
\description{
This function calculates the cosine similarity between the words in two specified columns of a dataframe using a provided word embedding model. It adds a new column, `cosine_similarity`, to the dataframe with the calculated similarity scores.
}
\details{
The function checks if the specified columns exist in the dataframe. For each pair of words (one from the `target_col` and one from the `response_col`) in each row, it retrieves their word embeddings using the provided model and computes the cosine similarity between them. If either word is invalid (e.g., contains spaces or commas), or if there's an error in processing, it returns `NA` for that pair. The resulting cosine similarity values are added to a new column named `cosine_similarity`.
}
