% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plural_singular.R
\name{is_plural_singular}
\alias{is_plural_singular}
\title{Check if Two Words are Singular and Plural Forms of Each Other}
\usage{
is_plural_singular(target_col, response_col)
}
\arguments{
\item{target_col}{A character string representing the first word.}

\item{response_col}{A character string representing the second word.}
}
\value{
A logical value (`TRUE` or `FALSE`). Returns `TRUE` if `target_col` and `response_col` are singular and plural forms of each other, otherwise `FALSE`.
}
\description{
Determines whether two given Spanish words are singular and plural forms of each other. The function evaluates common singular-plural transformations and handles special cases, including words ending in "s" or "x" and irregular plurals.
}
\details{
This function checks if the words are related by common singular-plural transformations. It handles various suffix changes and some irregular pluralization cases:
- Words ending in "s" or "x" are considered to not change their plural form.
- Handles irregular plurals, such as "luz" becoming "luces".
}
\examples{
is_plural_singular("canción", "canciones")  # TRUE
is_plural_singular("luz", "luces")          # TRUE
is_plural_singular("mono", "mona")          # FALSE

}
