% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_response_word.R
\name{is_response_word}
\alias{is_response_word}
\title{Check if Response is a Word from Specified Sources}
\usage{
is_response_word(
  dataframe,
  response_col,
  source1 = NULL,
  source2 = NULL,
  source3 = NULL
)
}
\arguments{
\item{dataframe}{A data frame containing the response data.}

\item{response_col}{A string specifying the name of the column in the dataframe that contains the responses to be checked.}

\item{source1}{An optional vector of words. This is the first source against which the responses will be checked. Defaults to `NULL`.}

\item{source2}{An optional vector of words. This is the second source against which the responses will be checked. Defaults to `NULL`.}

\item{source3}{An optional vector of words. This is the third source against which the responses will be checked. Defaults to `NULL`.}
}
\value{
A data frame with an additional column named `is_response_word`. This column is a numeric vector where `1` indicates that the response is a word found in the provided sources, and `0` indicates it is not.
}
\description{
This function checks if the values in the specified response column are found within any of the provided word source lists. It adds a new column, `is_response_word`, to the dataframe indicating whether each response is present in the combined source list.
}
\details{
The function checks if each value in the specified response column exists in any of the provided source lists (source1, source2, source3). The sources are first combined into a single list, excluding any `NULL` values, and then checked for duplicates to ensure uniqueness. The function then creates a new column, `is_response_word`, in the dataframe, where a value of `1` indicates the response is found in the combined sources, and `0` indicates it is not.
}
