% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_target_nonword.R
\name{is_target_nonword}
\alias{is_target_nonword}
\title{Check if the Item Type is a Nonword}
\usage{
is_target_nonword(item_type)
}
\arguments{
\item{item_type}{A character string representing the type of the item.}
}
\value{
A logical value (`TRUE` or `FALSE`). Returns `TRUE` if `item_type` matches any of the specified nonword variations, otherwise `FALSE`.
}
\description{
Determines whether a given item type represents a nonword by checking if it contains "nonword", "NW", "pseudoword", "pseudo-word", or "non-word".
}
\details{
This function checks if `item_type` matches common variations of the term "nonword", including "nonword", "NW", "pseudoword", "pseudo-word", and "non-word".
}
