% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_responses.R
\name{join_responses}
\alias{join_responses}
\title{Join several columns in only one response column}
\usage{
join_responses(
  df,
  resp_start = "response",
  resp_end = ncol(df),
  separator = ", ",
  drop_merged_cols = TRUE,
  remove_NA_values = TRUE,
  show_previous = TRUE
)
}
\arguments{
\item{df}{A data frame containing the responses to split.}

\item{resp_start}{The starting column for responses to join. This can be specified by column name (e.g., "response") or column position (e.g., 9). All response columns must be contiguous.}

\item{resp_end}{The ending column for responses to join. By default, this is the last column in the data frame (i.e., `ncol(df)`).}

\item{separator}{A character string used to separate the joined responses. Default is ", ".}

\item{drop_merged_cols}{Logical indicating whether to drop the original columns after merging. If `TRUE`, the original columns will be removed. Default is `TRUE`.}

\item{remove_NA_values}{Logical indicating whether to exclude `NA` values from the joined responses. If `TRUE`, `NA` values will be removed as rows. Default is `TRUE`.}

\item{show_previous}{Logical indicating whether to include the columns not used in the merge (e.g., identifier columns) in the output. If `TRUE`, these columns will be included. Default is `TRUE`.}
}
\value{
A data frame if `show_previous = TRUE`, or a tibble if `show_previous = FALSE`.
}
\description{
Joins several columns containing responses into a single column, with a specified separator between the responses.
}
