% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_scores.R
\name{position_scores}
\alias{position_scores}
\title{Split Strings into Columns and Transform to Long Format}
\usage{
position_scores(df, match_col, last_ID_col)
}
\arguments{
\item{df}{A data frame containing the column to be split and transformed.}

\item{match_col}{The name of the column in the data frame to be split into multiple columns.}

\item{last_ID_col}{The name of the last column in the original data frame that should be included in the final output. Columns after this will be split into separate columns and included in the long format.}
}
\value{
A data frame in long format with the original data and the new columns for character positions.
}
\description{
This function splits strings in a specified column of a data frame into individual characters, creates separate columns for each character position, and then transforms the data from wide to long format. Useful for analyzing character-level data or responses.
}
\examples{

data("IGC_long")

result <- get_formal_metrics(df = IGC_long,
                                item_col = "item",
                                response_col = "Response",
                                attempt_col = "Attempt",
                                group_cols = c("ID", "item_ID"))

result_positions <- result \%>\% position_scores(match_col = "itemL_adj_strict_match_pos",
                                              last_ID_col = "targetL")
print(result_positions)

}
