% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_letter_chart.R
\name{same_letter_start}
\alias{same_letter_start}
\title{Check if two words start with the same letter}
\usage{
same_letter_start(target_col, response_col)
}
\arguments{
\item{target_col}{A character string representing the first word.}

\item{response_col}{A character string representing the second word.}
}
\value{
A logical value (`TRUE` or `FALSE`). Returns `TRUE` if `target_col` and `response_col` start with the same letter, otherwise `FALSE`.
}
\description{
Determines whether two given words start with the same letter. It compares the first character of `target_col` and `response_col`.
}
\details{
This function extracts the first character of each string and checks if they are identical. It returns `TRUE` if both words start with the same letter, and `FALSE` otherwise.
}
\examples{
same_letter_start("monkey", "monye")
same_letter_start("article", "artim")

}
