% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same_root.R
\name{same_root}
\alias{same_root}
\title{Check if Two Words Share the Same Root}
\usage{
same_root(target_col, response_col)
}
\arguments{
\item{target_col}{A character string representing the first word.}

\item{response_col}{A character string representing the second word.}
}
\value{
A logical value (`TRUE` or `FALSE`). Returns `TRUE` if `target_col` and `response_col` share the same root. Returns `NA` if either word contains a comma or space.
}
\description{
Determines whether two words share the same root in Spanish using stemming. If the words contain a comma or space, the function returns `NA` to avoid processing multiple productions or concatenated words.
}
\details{
This function uses the `SnowballC` package to perform stemming on the `target_col` and `response_col` words. If either word contains a comma or space, the function skips the stemming process and returns `NA`. Otherwise, it compares the roots of the two words and returns `TRUE` if they match, or `FALSE` otherwise.
}
\examples{
same_root("hablando", "hablar")
same_root("corriendo", "correr")

}
