% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation_counting.R
\name{separation_counting}
\alias{separation_counting}
\title{Count the number of separations to determine the number of columns}
\usage{
separation_counting(
  df,
  col_name,
  separate_with = ", ",
  new_col_names = "attempt"
)
}
\arguments{
\item{df}{A data frame to work with.}

\item{col_name}{The target column containing the responses that will be analyzed.}

\item{separate_with}{A character string used to separate the responses (e.g., " - ", " ", or " & "). By default, it uses commas followed by an empty space (i.e., ", ").}

\item{new_col_names}{A base name for the new columns that will be created. The actual column names will be this base name followed by the column number (e.g., "attempt_1", "attempt_2"). Default is \code{"attempt"}.}
}
\value{
A list containing two elements: \code{column_names}, a character vector with the names of the new columns; and \code{count}, an integer representing the number of columns needed.
}
\description{
Calculate the number of occurrences of a specified separator character (e.g., ", ") in a response column to determine how many columns will be needed to separate the responses provided by a patient. It depends on the \code{df_from_column()} function and supports the \code{separate_responses()} function.
}
