% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_proportion_pos_ok.R
\name{shared_proportion_pos_ok}
\alias{shared_proportion_pos_ok}
\title{Calculate proportion of shared characters between two words in their correct position}
\usage{
shared_proportion_pos_ok(target_col, response_col)
}
\arguments{
\item{target_col}{A string representing the target word to compare.}

\item{response_col}{A string representing the response word to compare against the target.}
}
\value{
A numeric value representing the ratio of characters shared by the two strings as they are produced in the same position.
}
\description{
Compute the proportion of matching characters (letters/phonemes) between the words in `target_col` and `response_col`, where characters must be in the correct position. Specifically, this function compares each character in the target string to the corresponding character in the response string and calculates the proportion characters produced in the correct position.
}
\examples{
shared_proportion_pos_ok("monkey", "monye")
shared_proportion_pos_ok("article", "artim")

}
