% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_vector.R
\name{similarity_vector}
\alias{similarity_vector}
\title{Calculate the similarity vector between two words}
\usage{
similarity_vector(target_col, response_col)
}
\arguments{
\item{target_col}{A string representing the target word to compare.}

\item{response_col}{A string representing the response word to compare against the target.}
}
\value{
A numeric vector representing the similarity scores between characters in `target_col` and `response_col`. This vector reflects the degree of similarity based on character edits, with higher values indicating greater similarity.
}
\description{
Computes the similarity vector between the words located in `target_col` and `response_col`. This function uses the `adist` function from the `utils` package to calculate the similarity matrix and returns the vector of similarity scores by dropping transformation attributes.
}
\details{
The function utilizes the `adist` function to calculate a similarity matrix between the `target_col` and `response_col`. It then drops the transformation attributes from the result to obtain the similarity vector, which reflects the degree of similarity based on character edits.
}
\examples{
similarity_vector("monkey", "monye")
similarity_vector("article", "artim")

}
