% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Classify Responses Based on Multiple Criteria}
\usage{
classification(dataframe, access_col = "accessed", RA_col = "RA")
}
\arguments{
\item{dataframe}{A data frame containing the response data to be classified.}

\item{access_col}{A string specifying the name of the column in the dataframe that indicates whether the response was accessed (default is "accessed").}

\item{RA_col}{A string specifying the name of the column in the dataframe that indicates whether the response was a related attempt (default is "RA").}
}
\value{
A data frame with additional columns for each classification category: `lexicalization`, `nonword`, `neologism`, `formal`, `unrelated`, `morphological`, `mixed`, and `semantic`. Each column contains binary values indicating the classification status of each response.
}
\description{
This function classifies responses in a dataframe into different categories based on specified criteria. It adds new columns to the dataframe for each classification category, including lexicalization, nonword, neologism, formal, unrelated, morphological, mixed, and semantic.
}
\details{
The function processes the dataframe by renaming columns and applying a series of conditions to create new classification columns. The conditions are as follows:
- `lexicalization`: Identifies responses that are lexicalized based on specific conditions.
- `nonword`: Identifies responses that are classified as nonwords.
- `neologism`: Identifies responses that are classified as neologisms.
- `formal`: Identifies responses that are classified as formal.
- `unrelated`: Identifies responses that are classified as unrelated.
- `morphological`: Identifies responses that are classified as morphological.
- `mixed`: Identifies responses that are classified as mixed.
- `semantic`: Identifies responses that are classified as semantic.

Each condition uses logical operations to check the values in the dataframe's columns and applies criteria based on those values to determine the classification. The resulting classifications are added as new columns in the dataframe with binary indicators (1 for true, 0 for false).
}
