% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attempts.R
\name{get_attempts}
\alias{get_attempts}
\title{Obtain attempts and responses in long format}
\usage{
get_attempts(df, first_production = Attempt_1, drop_blank_spaces = FALSE)
}
\arguments{
\item{df}{A data frame containing the responses to split.}

\item{first_production}{The column from which attempts start. For example, if `first_production` is `Attempt_1`, it will include `Attempt_1` and all columns to its right as subsequent attempts. Default is `Attempt_1`.}

\item{drop_blank_spaces}{Logical. If `TRUE`, rows with empty responses are removed. If `FALSE`, these rows are kept in the output. Default is `FALSE`.}
}
\value{
A data frame in long format with columns `Attempt` and `Response`.
  - `Attempt`: Indicates the attempt number.
  - `Response`: Shows the response for the corresponding attempt, with `"#NONE"` replaced by `NA` if `drop_blank_spaces` is `TRUE`.
}
\description{
Transform a range of columns in a data frame from wide to long format. It starts from a specified column, referred to as `first_production`, and includes all columns to its right, which represent subsequent attempts. The function creates two new columns: `Attempt`, indicating the attempt number, and `Response`, showing the response for each attempt. Depending on the `drop_blank_spaces` parameter, rows with empty responses can be removed.
}
\examples{
# Example usage:
require(tidyverse)
example_data = data.frame(ID = 1:4,
  item = c("article", "firefighter", "honey", "money"),
  Response = c("ar, arm, art, article", "fire, far, firt, fire, firefigo, firefighter", "honey", "money"),
  stringsAsFactors = FALSE
)

df_attempts = get_attempts(df = example_data, first_production = resp_1, drop_blank_spaces = FALSE)

}
