% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_responses.R
\name{separate_responses}
\alias{separate_responses}
\title{Split the response in consecutive columns}
\usage{
separate_responses(df, col_name, separate_with = ", ", show_previous = TRUE)
}
\arguments{
\item{df}{A data frame containing the responses to join together.}

\item{col_name}{Target column to be considered as response column.}

\item{separate_with}{A character string used separate the responses (e.g., " - ", " ", or " & "). By default it uses commas (i.e., ", ").}

\item{show_previous}{Logical; whether the previous columns not used in the merge (i.e. the identifier columns) are provided in the output or not. If \code{'FALSE'} only the separated columns are provided in the output. Default is \code{'TRUE'}.}
}
\value{
A data frame.
}
\description{
Separates the text of an answer column into consecutive independent columns according to the number of separations in that answer column. Depends on the \code{counting_to_split()} function.
}
\details{
This function relies on `separation_counting()`.
}
\examples{

data("IGC")

IGC \%>\% separate_responses(col_name = "final_response",
                                separate_with = ", ",
                                show_previous = F)
separate_responses(df = IGC, col_name = "final_response",
                                separate_with = ", ",
                                show_previous = F)
}
